Split server bootstrap
