+++ twake/backend/node/src/server.ts
@@ -1,7 +1,7 @@
 import * as Sentry from "@sentry/node";
-import path from "path";
-import { TwakePlatform, TwakePlatformConfiguration } from "./core/platform/platform";
+import { TwakePlatform } from "./core/platform/platform";
 import config from "./core/config";
+import twake from "./twake";
 
 if (config.get("sentry.dsn")) {
   Sentry.init({
@@ -10,46 +10,6 @@ if (config.get("sentry.dsn")) {
   });
 }
 
-let platform: TwakePlatform;
+const launch = async (): Promise<TwakePlatform> => twake.run(config.get("services"));
 
-const start = async (): Promise<TwakePlatform> => {
-  try {
-    const configuration: TwakePlatformConfiguration = {
-      services: config.get("services"),
-      servicesPath: path.resolve(__dirname, "./services/"),
-    };
-    platform = new TwakePlatform(configuration);
-    await platform.init();
-    await platform.start();
-
-    return platform;
-  } catch (err) {
-    console.error(err);
-    process.exit(-1);
-  }
-};
-
-process.on("uncaughtException", error => {
-  console.error(error);
-});
-
-process.on("unhandledRejection", error => {
-  console.error(error);
-});
-
-process.on("SIGINT", stop);
-
-process.once("SIGUSR2", async () => {
-  await stop();
-  process.kill(process.pid, "SIGUSR2");
-});
-
-async function stop() {
-  try {
-    await platform?.stop();
-  } catch (err) {
-    console.error(err);
-  }
-}
-
-start();
+launch();
